move GroupByClause to pkg/ast
Moves the GroupByClause struct to `pkg/ast` and adds a
`pkg/ast.NewGroupByClause` function for external callers.
